+++ example/views/DragAndDropWithTouchSupportMultipleTreesExample.vue
@@ -9,6 +9,7 @@
         :enableDragNodeOut="true"
         :enableDropExternalElement="true"
         :enableTouchSupport="true"
+        :allowedTouchFromTreeIds="['rightTree']"
         @drop="drop('leftTree', arguments[0])"/>
       <TWTree
         ref="rightTree" 
@@ -18,6 +19,7 @@
         :enableDragNodeOut="true"
         :enableDropExternalElement="true"
         :enableTouchSupport="true"
+        :allowedTouchFromTreeIds="['leftTree']"
         @drop="drop('rightTree', arguments[0])"/>
     </div>
   </div>
@@ -169,10 +171,6 @@ export default {
           break
       }
     }
-  },
-  mounted () {
-    this.$refs.leftTree.allowTouchOperationFromAnotherTree('rightTree')
-    this.$refs.rightTree.allowTouchOperationFromAnotherTree('leftTree')
   }
 }
 </script>